feat(init): initial setup and configuration        Added a server.js file for setting up the routes and ports for the application.        Added a prettierrc.json file to make all the quotes single.        Added a prettierignore file to ignore the .md files.        Added an eslintrc.json to change the configuration of the application so that it's easier to work on the code.        Added an eslintignore file to ignore node files.        Added an env file to hide sensitive information.        Added a connection.js file to the config for connecting to the server to the database.        Added an index.js file to the controllers folder for handling the controller functions.        Added an api folder in the controllers folder to include an index.js file for the api routes that are going to be exported to help manage where the routes are going.        Added the schema.sql file to the db folder for the database.        Added an index.js file to the models folder for handling what the view does in the MVC template being used.        Added a seed.js file to the seeds folder to handle seeding the server's database.
feat(init): initial setup and configuration

Added a server.js file for setting up the routes and ports for the application.

Added a prettierrc.json file to make all the quotes single.

Added a prettierignore file to ignore the .md files.

Added an eslintrc.json to change the configuration of the application so that it's easier to work on the code.

Added an eslintignore file to ignore node files.

Added an env file to hide sensitive information.

Added a connection.js file to the config for connecting to the server to the database.

Added an index.js file to the controllers folder for handling the controller functions.

Added an api folder in the controllers folder to include an index.js file for the api routes that are going to be exported to help manage where the routes are going.

Added the schema.sql file to the db folder for the database.

Added an index.js file to the models folder for handling what the view does in the MVC template being used.

Added a seed.js file to the seeds folder to handle seeding the server's database.