feat(server): set up server with Express and Sequelize.        Added code to the server that sets up a server using express and sequelize.        Added code to include the required modules such as path, express, express-session, express-handlebars, and connect-session-sequelize.        Added code to import the routes from controllers.        Added middleware to initialize the Express application to set the port number to the env variable PORT or 3001.        Added the middleware to configure the session with a secret key, an empty cookie object, and established sequelize to deal with the objects.        Included code to set up the express application to use sessions and handlebars as the view engine.        Added code to configure the app to parse JSON and URL-encoded data, serve static files from the 'public' directory, and use the defined routes.        Added code to synchronize the sequelize models with the database and included code that will allow the server to listen on the specified port.
feat(server): set up server with Express and Sequelize

Added code to the server that sets up a server using express and sequelize.

Added code to include the required modules such as path, express, express-session, express-handlebars, and connect-session-sequelize.

Added code to import the routes from controllers.

Added middleware to initialize the Express application to set the port number to the env variable PORT or 3001.

Added the middleware to configure the session with a secret key, an empty cookie object, and established sequelize to deal with the objects.

Included code to set up the express application to use sessions and handlebars as the view engine.
Added code to configure the app to parse JSON and URL-encoded data, serve static files from the 'public' directory, and use the defined routes.

Added code to synchronize the sequelize models with the database and included code that will allow the server to listen on the specified port.