feat(public): add styling and dynamic functionality for login/logout        Added a jass.css file to the public directory's css folder for the styling of the website.        Added a login.js and logout.js to the public directory's js folder to add dynamics to the login and logout forms on the website.
feat(public): add styling and dynamic functionality for login/logout

Added a jass.css file to the public directory's css folder for the styling of the website.

Added a login.js and logout.js to the public directory's js folder to add dynamics to the login and logout forms on the website.