feat(controller): add editPost function for handling post editing  Added the editPost function with an event listener to handle editing a post by ID. If a title and content value exist, a PUT request is made. If the response is ok, the user is redirected to the dashboard; otherwise, an alert is shown.
feat(controller): add editPost function for handling post editing

Added the editPost function with an event listener to handle editing a post by ID. If a title and content value exist, a PUT request is made. If the response is ok, the user is redirected to the dashboard; otherwise, an alert is shown.