fix(views): Fix routing for sign in and sign up buttons  Updated the scripts in the views folder to handle routing for the sign in and sign up buttons. Added IDs to inputs for the scripts included in the HTML of the Handlebars files. Updated routes to render pages when they are loaded.
fix(views): Fix routing for sign in and sign up buttons

Updated the scripts in the views folder to handle routing for the sign in and sign up buttons.
Added IDs to inputs for the scripts included in the HTML of the Handlebars files.
Updated routes to render pages when they are loaded.