feat(server): add session expiration and logging out, comments, and model updates  Added cookies to the server file to log out a user if the session expires so they are prompted to log back in when they revisit and idle for more than a minute.   Added comments to files to add clarity. Removed date from the post model.   Added hooks true to the models index file so that lifecycles for these relations will be read properly.
feat(server): add session expiration and logging out, comments, and model updates

Added cookies to the server file to log out a user if the session expires so they are prompted to log back in when they revisit and idle for more than a minute.

Added comments to files to add clarity. Removed date from the post model.

Added hooks true to the models index file so that lifecycles for these relations will be read properly.