feat(models): update postdata.json and add comment functionality  Changed the postdata.json to include relevant data for testing.   Added an "Add Comment" button to the homepage.handlebars for users that are logged in.   Added user_id to the comments and users models, and post_id to the post model so they can be referenced correctly when adding, updating, and removing objects.   Added viewPost and editPost routes to the postRoutes file (not working yet, will fix).   Updated the login.js file to alert the user when the login has failed (not working yet, will fix).
feat(models): update postdata.json and add comment functionality

Changed the postdata.json to include relevant data for testing.

Added an "Add Comment" button to the homepage.handlebars for users that are logged in.

Added user_id to the comments and users models, and post_id to the post model so they can be referenced correctly when adding, updating, and removing objects.

Added viewPost and editPost routes to the postRoutes file (not working yet, will fix).

Updated the login.js file to alert the user when the login has failed (not working yet, will fix).