feat(styles): enhance user interface with new styles  Included styles in the style.css to create a polished application that is more user friendly. The updated styles provide a consistent and visually appealing interface across the login and signup forms.  Included the post title and post content in the edit post display so that the user knows which post they are editing. This ensures clarity and improves the editing experience.  Updated images for the website to enhance visual content and overall aesthetics.
feat(styles): enhance user interface with new styles

Included styles in the style.css to create a polished application that is more user friendly. The updated styles provide a consistent and visually appealing interface across the login and signup forms.

Included the post title and post content in the edit post display so that the user knows which post they are editing. This ensures clarity and improves the editing experience.

Updated images for the website to enhance visual content and overall aesthetics.